requirements.txt
llama-index
openai
pypdf
python-dotenv
llama-index-llms-gemini
llama-index-embeddings-gemini